Refactor rocket_utils into the scpui system functions submodel. Rename rocket_templates to templates to match other uility files for now.
